move proto classes
